Edited forms for registration, password reset, add vinyl. Edited api view for LP delete. Added images urls for manually added LP & not founded LP.

+++ api/views.py
@@ -1,7 +1,7 @@
 from rest_framework import generics
 from rest_framework.authentication import SessionAuthentication
 from rest_framework.decorators import authentication_classes
-from rest_framework.permissions import IsAuthenticated
+from rest_framework.permissions import IsAuthenticated, BasePermission
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from vinylcollector.models import Vinyl, User
@@ -27,7 +27,7 @@ class RegisterView(APIView):
 
 
 class VinylListView(APIView):
-    # permission_classes = (IsAuthenticated,)
+    permission_classes = (IsAuthenticated,)
 
     @staticmethod
     def get(request):
@@ -68,8 +68,12 @@ class VinylDetailsView(APIView):
     @staticmethod
     def delete(request, pk):
         vinyl = generics.get_object_or_404(Vinyl, pk=pk)
-        vinyl.delete()
+        if vinyl.owner == request.user:
+            vinyl.delete()
+            return Response({
+                "message": f"Vinyl with catalogue number:'{vinyl.catalogue_number}' has been deleted."
+            },)
         return Response({
-            "message": f"Vinyl with catalogue number:'{vinyl.catalogue_number}' has been deleted."
-        },)
+            "message": f"Vinyl is not in yor collection"
+        }, )
